Move datatype strategies into datatype::strategy module
