+++ tiledb/api/src/strategy/filter.rs
@@ -3,9 +3,9 @@ use std::collections::VecDeque;
 use proptest::prelude::*;
 use proptest::strategy::Just;
 
+use crate::datatype::strategy::*;
 use crate::filter::*;
 use crate::filter_list::FilterListData;
-use crate::strategy::datatype::*;
 use crate::Datatype;
 
 fn prop_bitwidthreduction() -> impl Strategy<Value = FilterData> {
