+++ tiledb/api/src/strategy/mod.rs
@@ -1,5 +1,3 @@
-pub mod datatype;
 pub mod filter;
 
-pub use datatype::*;
 pub use filter::*;
